+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG.md
 
-## 2.1.5 (2025-04-06)
+## 2.2.1 (2025-04-06)
+
+- Removed `anchor` and `portal` props from `TypeaheadSelect` to fix dropdown menu behavior.
+- Fixed styling when using arrow keys for navigation in `TypeaheadSelect` component.
+
+## 2.2.0 (2025-04-06)
 
 - Removed default value for `anchor` prop in `Select`, `DropdownMenu`, and `Popover` components.
 - Added `anchor` prop to `TypeaheadSelect` component.